third_party/blink/web_tests/crypto/subtle/importKey-badParameters-expected.txt

CONSOLE WARNING: line 68: SharedArrayBuffer will require cross-origin isolation as of M91, around May 2021. See https://developer.chrome.com/blog/enabling-shared-array-buffer/ for more details.
Tests calling cypto.subtle.importKey with bad parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

error is: TypeError: Key data must be a BufferSource for non-JWK formats
error is: TypeError: Key data must be a BufferSource for non-JWK formats
error is: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': Failed to convert value to 'object'.
error is: TypeError: Invalid keyFormat argument
error is: TypeError: Invalid keyUsages argument
error is: TypeError: Invalid keyFormat argument
error is: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': HmacImportParams: hash: Missing or not an AlgorithmIdentifier
error is: NotSupportedError: Failed to execute 'importKey' on 'SubtleCrypto': SHA-1: Unsupported operation: importKey
error is: TypeError: Failed to execute 'importKey' on 'SubtleCrypto': The provided ArrayBufferView value must not be shared.
PASS successfullyParsed is true

TEST COMPLETE
